+++ app/frontend/src/javascript/api/proof-of-identity-file.ts
@@ -1,10 +1,11 @@
 import apiClient from './clients/api-client';
 import { AxiosResponse } from 'axios';
 import { ProofOfIdentityFile, ProofOfIdentityFileIndexFilter } from '../models/proof-of-identity-file';
+import ApiLib from '../lib/api';
 
 export default class ProofOfIdentityFileAPI {
   static async index (filters?: ProofOfIdentityFileIndexFilter): Promise<Array<ProofOfIdentityFile>> {
-    const res: AxiosResponse<Array<ProofOfIdentityFile>> = await apiClient.get(`/api/proof_of_identity_files${this.filtersToQuery(filters)}`);
+    const res: AxiosResponse<Array<ProofOfIdentityFile>> = await apiClient.get(`/api/proof_of_identity_files${ApiLib.filtersToQuery(filters)}`);
     return res?.data;
   }
 
@@ -26,11 +27,5 @@ export default class ProofOfIdentityFileAPI {
   static async destroy (proofOfIdentityFileId: number): Promise<void> {
     const res: AxiosResponse<void> = await apiClient.delete(`/api/proof_of_identity_files/${proofOfIdentityFileId}`);
     return res?.data;
-  }
-
-  private static filtersToQuery (filters?: ProofOfIdentityFileIndexFilter): string {
-    if (!filters) return '';
-
-    return '?' + Object.entries(filters).map(f => `${f[0]}=${f[1]}`).join('&');
   }
 }