+++ app/frontend/src/javascript/models/api.ts
@@ -1,0 +1,3 @@
+// ApiFilter should be extended by an interface listing all the filters allowed for a given API
+// eslint-disable-next-line @typescript-eslint/ban-types
+export type ApiFilter = {};