+++ app/frontend/src/javascript/models/group.ts
@@ -1,4 +1,6 @@
-export interface GroupIndexFilter {
+import { ApiFilter } from './api';
+
+export interface GroupIndexFilter extends ApiFilter {
   disabled?: boolean,
   admins?: boolean,
 }