+++ app/policies/credit_policy.rb
@@ -1,3 +1,6 @@
+# frozen_string_literal: true
+
+# Check the access policies for API::CreditsController
 class CreditPolicy < ApplicationPolicy
   def index?
     user.admin?
@@ -13,5 +16,9 @@ class CreditPolicy < ApplicationPolicy
 
   def destroy?
     index?
+  end
+
+  def user_resource?
+    record.id == user.id
   end
 end