+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## next deploy
 
+## v5.4.8 2022 June 29
+
+- My reservations dashboard
+- Display reservations credits in the dashboard
+- Added a test case for space reservation
+- Updated Portugueses translations (thanks to [@ghabreu](https://github.com/ghabreu))
+- Improved explanations about members CSV imports
+- Fix a bug: unable to reserve a space
+- Fix a bug: invalid users are reported in search
+
 ## v5.4.7 2022 June 29
 
 - Removed the admins' button to scroll to the featured event